requirements.txt
fastapi
uvicorn
openai-whisper
requests
python-dotenv